\name{load454SampleData}
\alias{load454SampleData}

\title{Load the example 454 dataset}

\description{
Load a small, 1000 record example sff file.
}

\usage{
load454SampleData()
}

\arguments{
}

\value{A \code{\linkS4class{SffReadsQ}} object containing 1000 reads
with lengths ranging from 42 to 952 basepairs.}

\author{Matt Settles <msettles@uidaho.edu>}
