\name{loadIonSampleData}
\alias{loadIonSampleData}

\title{Load the example Ion Torrent dataset}

\description{
Load a small, 1000 record example sff file.
}

\usage{
load454SampleData()
}

\arguments{
}

\value{A \code{\linkS4class{SffReadsQ}} object containing 1000 reads
with lengths ranging from 5 to 353 basepairs.}

\author{Matt Settles <msettles@uidaho.edu>}
