\name{rSFFreader internals}
\docType{class}

% .sffValidity class
\alias{class:.sffValidity}
\alias{.sffValidity-class}
\alias{.sffValidity}

% .SFFBase class and methods:
\alias{class:.SFFBase}
\alias{.SFFBase-class}
\alias{.SFFBase}

\alias{detail,.SFFBase-method}
\alias{show,.SFFBase-method}


\title{rSFFreader internals}

\description{
  rSFFreader objects, classes and methods that are not intended to
  be used directly.
}

\author{M. Settles}

\keyword{internal}
\keyword{methods}
\keyword{classes}
