\name{rSFFreader-package}
\alias{rSFFreader-package}
\alias{rSFFreader}
\alias{.SFFBase-class}
\alias{show,.SFFBase-method}
\alias{append,.SFFBase,.SFFBase,ANY-method}
\docType{package}
\title{

  Base classes and methods for high-throughput sequence data stored in SFF files.

}
\description{

  Base classes, functions, and methods for representation of
  high-throughput sequencing data stored in SFF files. Such as 
  Roche 454 Data and Life Sciences Ion Torrent

}
\details{

  See \code{packageDescription('rSFFreader')}

}
\author{
Matt Settles <msettles@uidaho.edu> and Kyu-chul Cho
Maintainer: Matt Settles <msettles@uidaho.edu>
}
\references{
There will be one
}
\keyword{ package }
