\name{load454SampleData}
\alias{load454SampleData}

\title{Load the example 454 dataset}

\description{
Load a small, 1000 record example sff file.
}

\usage{
load454SampleData()
}

\arguments{
None
}

\value{Loads a small example Roche 454 dataset(sff file) into a \code{\linkS4class{SffReadsQ}} object containing 1000 reads.}

\author{Matt Settles <msettles@uidaho.edu>}
