\name{loadIonSampleData}
\alias{loadIonSampleData}

\title{Load the example Ion Torrent dataset}

\description{
Load a small, 1000 record example sff file.
}

\usage{
load454SampleData()
}

\arguments{
None
}

\value{Loads a small example Life Sciences Ion Torrent dataset (sff file) into a \code{\linkS4class{SffReadsQ}} object containing 1000 reads.}

\author{Matt Settles <msettles@uidaho.edu>}
