\name{rSFFreader internals}
\docType{class}

% .sffValidity class
\alias{.sffValidity}
\alias{.solveIRangeSEW}

\title{rSFFreader internals}

\description{
  rSFFreader objects, classes and methods that are not intended to
  be used directly.
}

\author{M. Settles}

\keyword{internal}
\keyword{methods}
\keyword{classes}
