\name{rSFFreader-package}
\alias{rSFFreader-package}
\alias{rSFFreader}
\docType{package}

\title{

  Base classes and methods for high-throughput sequence data stored in SFF files.

}
\description{

  Base classes, functions, and methods for representation of
  high-throughput sequencing data stored in SFF files (such as 
  Roche 454 Data and Life Sciences Ion Torrent). Implementation defined
  according to http://www.ncbi.nlm.nih.gov/Traces/trace.cgi?cmd=show&f=formats&m=doc&s=formats#sff
  

}
\details{

  See \code{packageDescription('rSFFreader')}

}
\author{
Matt Settles <msettles@uidaho.edu> and Kyu-chul Cho
Maintainer: Matt Settles <msettles@uidaho.edu>
}
\references{
Coming Soon
}
\keyword{ package }
